feat: Add function size() to singly linked list
This commit adds a function size() to our C++ implementation of a singly linked list. The size() function returns the number of nodes in the linked list.

In the main function, I have added a case to show the size value of an example of a singly linked list.

This new feature will make it easier to use our singly linked list implementation, as it will allow us to easily determine the size of the linked list.